requirements.txt
flask==3.0.0
openai==1.52.0
python-dotenv==1.0.0
pandas==2.1.4
pillow==10.1.0
gunicorn==21.2.0
